requirements.txt
tensorflow==2.18.0
mediapipe==0.10.14
streamlit==1.39.0
opencv-python==4.8.0.76
numpy==1.26.4
gtts==2.5.3
rich==13.7.1
nltk==3.8.1
requests==2.31.0
streamlit-camera-input-live==0.2.0
Pillow==9.5.0
h5py==3.14.0